Revert "factor out identity test code"
This reverts commit 787a117906d75430ef3ab254e9f18bfe91ddc521.

Turns out we actually no longer need any of that!
