+++ linkerd/app/integration/src/data/bar-ns1/token.txt
