+++ linkerd/app/integration/src/data/foo-ns1/ca1-cert.csr
