+++ linkerd/app/integration/src/identity.rs
@@ -1,12 +1,12 @@
 use super::*;
 use std::{
     collections::VecDeque,
-    path::PathBuf,
+    fs, io,
+    path::{Path, PathBuf},
     sync::{Arc, Mutex},
     time::{Duration, SystemTime},
 };
 
-use linkerd2_app_test::identity as app_test;
 use linkerd2_proxy_api::identity as pb;
 use tonic as grpc;
 
@@ -33,40 +33,129 @@ type Certify = Box<
         > + Send,
 >;
 
+const TLS_VERSIONS: &[rustls::ProtocolVersion] = &[rustls::ProtocolVersion::TLSv1_2];
+
+struct Certificates {
+    pub leaf: Vec<u8>,
+    pub intermediates: Vec<Vec<u8>>,
+}
+
+impl Certificates {
+    pub fn load<P>(p: P) -> Result<Certificates, io::Error>
+    where
+        P: AsRef<Path>,
+    {
+        let f = fs::File::open(p)?;
+        let mut r = io::BufReader::new(f);
+        let certs = rustls::internal::pemfile::certs(&mut r)
+            .map_err(|_| io::Error::new(io::ErrorKind::Other, "rustls error reading certs"))?;
+        let leaf = certs
+            .get(0)
+            .ok_or_else(|| io::Error::new(io::ErrorKind::Other, "no certs in pemfile"))?
+            .as_ref()
+            .into();
+        let intermediates = certs[1..].iter().map(|i| i.as_ref().into()).collect();
+
+        Ok(Certificates {
+            leaf,
+            intermediates,
+        })
+    }
+
+    pub fn chain(&self) -> Vec<rustls::Certificate> {
+        let mut chain = Vec::with_capacity(self.intermediates.len() + 1);
+        chain.push(self.leaf.clone());
+        chain.extend(self.intermediates.clone());
+        chain.into_iter().map(rustls::Certificate).collect()
+    }
+
+    pub fn response(&self) -> pb::CertifyResponse {
+        pb::CertifyResponse {
+            leaf_certificate: self.leaf.clone(),
+            intermediate_certificates: self.intermediates.clone(),
+            ..Default::default()
+        }
+    }
+}
+
 impl Identity {
+    fn load_key<P>(p: P) -> rustls::PrivateKey
+    where
+        P: AsRef<Path>,
+    {
+        let p8 = fs::read(&p).expect("read key");
+        rustls::PrivateKey(p8)
+    }
+
+    fn configs(
+        trust_anchors: &str,
+        certs: &Certificates,
+        key: rustls::PrivateKey,
+    ) -> (Arc<rustls::ClientConfig>, Arc<rustls::ServerConfig>) {
+        use std::io::Cursor;
+        let mut roots = rustls::RootCertStore::empty();
+        roots
+            .add_pem_file(&mut Cursor::new(trust_anchors))
+            .expect("add pem file");
+
+        let mut client_config = rustls::ClientConfig::new();
+        client_config.root_store = roots;
+
+        let mut server_config = rustls::ServerConfig::new(
+            rustls::AllowAnyAnonymousOrAuthenticatedClient::new(client_config.root_store.clone()),
+        );
+
+        server_config.versions = TLS_VERSIONS.to_vec();
+        server_config
+            .set_single_cert(certs.chain(), key)
+            .expect("set server resover");
+
+        (Arc::new(client_config), Arc::new(server_config))
+    }
+
     pub fn new(dir: &'static str, local_name: String) -> Self {
-        let app_test::Identity {
-            id_dir,
-            token_path,
-            trust_anchors,
-            certs:
-                app_test::Certificates {
-                    leaf,
-                    intermediates,
-                    ..
-                },
-            client_config,
-            server_config,
-            ..
-        } = app_test::Identity::new(dir);
-        let path_to_string = |path: PathBuf| path.into_os_string().into_string().unwrap();
+        let (id_dir, token, trust_anchors, certs, key) = {
+            let path_to_string = |path: &PathBuf| {
+                path.as_path()
+                    .to_owned()
+                    .into_os_string()
+                    .into_string()
+                    .unwrap()
+            };
+            let mut id = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
+            id.push("src");
+            id.push("data");
+
+            id.push("ca1.pem");
+            let trust_anchors = fs::read_to_string(&id).expect("read trust anchors");
+
+            id.set_file_name(dir);
+            let id_dir = path_to_string(&id);
+
+            id.push("token.txt");
+            let token = path_to_string(&id);
+
+            id.set_file_name("ca1-cert.pem");
+            let certs = Certificates::load(&id).expect("read cert");
+
+            id.set_file_name("key.p8");
+            let key = Identity::load_key(&id);
+
+            (id_dir, token, trust_anchors, certs, key)
+        };
+
+        let certify_rsp = certs.response();
+        let (client_config, server_config) = Identity::configs(&trust_anchors, &certs, key);
         let mut env = TestEnv::new();
 
-        env.put(app::env::ENV_IDENTITY_DIR, path_to_string(id_dir));
-        env.put(
-            app::env::ENV_IDENTITY_TOKEN_FILE,
-            path_to_string(token_path),
-        );
+        env.put(app::env::ENV_IDENTITY_DIR, id_dir);
+        env.put(app::env::ENV_IDENTITY_TOKEN_FILE, token);
         env.put(app::env::ENV_IDENTITY_TRUST_ANCHORS, trust_anchors);
         env.put(app::env::ENV_IDENTITY_IDENTITY_LOCAL_NAME, local_name);
 
         Self {
             env,
-            certify_rsp: pb::CertifyResponse {
-                leaf_certificate: leaf,
-                intermediate_certificates: intermediates,
-                ..Default::default()
-            },
+            certify_rsp,
             client_config,
             server_config,
         }
