+++ linkerd/app/test/src/lib.rs
@@ -17,7 +17,6 @@ use std::fmt;
 pub use tokio_test::io;
 
 pub mod connect;
-pub mod identity;
 pub mod refine;
 pub mod resolver;
 
@@ -34,10 +33,6 @@ pub fn connect<E: fmt::Debug>() -> connect::Connect<E> {
 
 pub fn io() -> io::Builder {
     io::Builder::new()
-}
-
-pub fn identity(dir: &'static str) -> identity::Identity {
-    identity::Identity::new(dir)
 }
 
 /// By default, disable logging in modules that are expected to error in tests.
